fix some things, abstract to library crate
