+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rust-gc-count"
+name = "rust_gc_count"
 version = "0.1.0"
 edition = "2021"
 authors = ["Andy Yates"]
