+++ src/checksumseq.rs
@@ -1,13 +1,12 @@
-use base64_url;
 use clap::Parser;
 use flate2::read::MultiGzDecoder;
-use md5::{Digest, Md5};
-use seq_io::fasta::{Reader, Record, RefRecord};
-use sha2::Sha512;
+use seq_io::fasta::Reader;
 use std::fs::File;
 use std::io::prelude::{Read, Write};
 use std::io::BufWriter;
 use std::io::{stdin, stdout};
+
+use rust_gc_count::checksum::process_sequence;
 
 /// Parse a FASTA file and calculate checksums for each record
 #[derive(Parser)]
@@ -65,7 +64,7 @@ fn main() {
             result.0, result.1, result.2, result.3
         );
         writer
-            .write(line.as_bytes())
+            .write_all(line.as_bytes())
             .expect("Could not write to file");
         if args.verbose {
             eprintln!("done");
@@ -76,25 +75,6 @@ fn main() {
         eprintln!("==> Found and processed {} regions.", n);
     }
     writer.flush().expect("Could not flush writer");
-}
-
-fn process_sequence(record: RefRecord, verbose: bool) -> (String, usize, String, String) {
-    let mut md5_hasher_box = Box::new(Md5::new());
-    let mut sha512_hasher_box = Box::new(Sha512::new());
-    let id = record.id().unwrap();
-    let mut length = 0;
-    if verbose {
-        eprint!("==> Processing region {:?} ... ", id);
-    }
-    for s in record.seq_lines() {
-        sha512_hasher_box.as_mut().update(s.to_ascii_uppercase());
-        md5_hasher_box.as_mut().update(s.to_ascii_uppercase());
-        length += s.len();
-    }
-    let sha512 = base64_url::encode(&sha512_hasher_box.as_mut().finalize_reset()[0..24]);
-    let md5 = format!("{:x}", md5_hasher_box.as_mut().finalize_reset());
-    
-    (id.to_string(), length, sha512, md5)
 }
 
 #[test]
